contrib/google.golang.org: creates tag_utils file to avoid repetitive use of helper method
Moves helper method ExtractRPCTags to tag_utils.go under internal/grpcutil
Adds test for tag_utils.go
Removes redundancies from grpc and grpc.v12 including redundant tests
