+++ contrib/google.golang.org/grpc/grpc.go
@@ -10,14 +10,12 @@ package grpc // import "gopkg.in/DataDog/dd-trace-go.v1/contrib/google.golang.or
 
 import (
 	"errors"
-	"io"
-	"math"
-	"strings"
-
 	"gopkg.in/DataDog/dd-trace-go.v1/contrib/google.golang.org/internal/grpcutil"
 	"gopkg.in/DataDog/dd-trace-go.v1/ddtrace"
 	"gopkg.in/DataDog/dd-trace-go.v1/ddtrace/ext"
 	"gopkg.in/DataDog/dd-trace-go.v1/ddtrace/tracer"
+	"io"
+	"math"
 
 	context "golang.org/x/net/context"
 	"google.golang.org/grpc/codes"
@@ -47,7 +45,7 @@ func (cfg *config) startSpanOptions(opts ...tracer.StartSpanOption) []tracer.Sta
 func startSpanFromContext(
 	ctx context.Context, method, operation, service string, opts ...tracer.StartSpanOption,
 ) (ddtrace.Span, context.Context) {
-	rpcTags := extractRPCTags(method)
+	rpcTags := grpcutil.ExtractRPCTags(method)
 	opts = append(opts,
 		tracer.ServiceName(service),
 		tracer.ResourceName(method),
@@ -86,34 +84,4 @@ func finishWithError(span ddtrace.Span, err error, cfg *config) {
 		}
 	}
 	span.Finish(finishOptions...)
-}
-
-// extractRPCTags will assign the proper tag values for method, service, package according to otel given a full method
-func extractRPCTags(fullMethod string) map[string]string {
-
-	//Otel definition: https://opentelemetry.io/docs/reference/specification/trace/semantic_conventions/rpc/#span-name
-
-	tags := map[string]string{
-		ext.RPCMethod:   "",
-		ext.RPCService:  "",
-		ext.GRPCPackage: "",
-	}
-
-	//Always remove leading slash
-	fullMethod = strings.TrimPrefix(fullMethod, "/")
-
-	//Split by slash and get everything after last slash as method
-	slashSplit := strings.SplitAfter(fullMethod, "/")
-	tags[ext.RPCMethod] = slashSplit[len(slashSplit)-1]
-
-	//Join everything before last slash and remove last slash as service
-	tags[ext.RPCService] = strings.TrimSuffix(strings.Join(slashSplit[:len(slashSplit)-1], ""), "/")
-
-	//Split by period and see if package exists if period is found
-	if strings.Contains(tags[ext.RPCService], ".") {
-		dotSplit := strings.SplitAfter(tags[ext.RPCService], ".")
-		tags[ext.GRPCPackage] = strings.TrimSuffix(strings.Join(dotSplit[:len(dotSplit)-1], ""), ".")
-	}
-
-	return tags
 }
