+++ contrib/google.golang.org/grpc/grpc_test.go
@@ -1047,54 +1047,3 @@ func BenchmarkUnaryServerInterceptor(b *testing.B) {
 		}
 	})
 }
-
-func TestExtractTags(t *testing.T) {
-	testCases := []struct {
-		Name    string
-		Path    string
-		Service string
-		Method  string
-		Package string
-	}{
-		{"empty test",
-			"",
-			"",
-			"",
-			"",
-		},
-		{"basic test",
-			"/mypackage.myservice/mymethod",
-			"mypackage.myservice",
-			"mymethod",
-			"mypackage",
-		},
-		{"obscure test",
-			"/my/p/a/c.k.a.ge.my/se/r/v/ice/myme.t.h.od",
-			"my/p/a/c.k.a.ge.my/se/r/v/ice",
-			"myme.t.h.od",
-			"my/p/a/c.k.a.ge",
-		},
-	}
-	for _, tc := range testCases {
-		t.Run(tc.Name, func(t *testing.T) {
-			tags := extractRPCTags(tc.Path)
-			if tags[ext.RPCService] != tc.Service {
-				t.Errorf("Service %s != %s", tags[ext.RPCService], tc.Service)
-			} else {
-				t.Logf("Service %s == %s", tags[ext.RPCService], tc.Service)
-			}
-
-			if tags[ext.RPCMethod] != tc.Method {
-				t.Errorf("Method %s != %s", tags[ext.RPCMethod], tc.Method)
-			} else {
-				t.Logf("Method %s == %s", tags[ext.RPCMethod], tc.Method)
-			}
-
-			if tags[ext.GRPCPackage] != tc.Package {
-				t.Errorf("Package %s != %s", tags[ext.RPCService], tc.Package)
-			} else {
-				t.Logf("Package %s == %s", tags[ext.RPCService], tc.Package)
-			}
-		})
-	}
-}
